requirements.txt
Django==5.2
mysqlclient
python-dotenv
gunicorn
boto3
django-storages
